pyproject.toml
[project]
name = "ishare-uploader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "mammoth>=1.11.0",
    "markdown>=3.10",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.53.0",
]
